use the keywords api for common lisp
